[NO GBP] Fixes polymorphing into drones gibbing you (#80527)
Uhh something about moving the contents from the human into the drones
hands, including your limbs and organs

:cl:
fix: polymorphing into drone no longer gibs
/:cl:
# Conflicts:
#	code/modules/mob/living/living.dm

+++ code/modules/mob/living/living.dm
@@ -1432,7 +1432,8 @@
 	else
 		for(var/obj/item/item in src)
 			if(!dropItemToGround(item))
-				qdel(item)
+				if(!(item.item_flags & ABSTRACT))
+					qdel(item)
 				continue
 			item_contents += item
